Refactor Nav to Include Graphviz Functionality for Diagram Display
This commit enhances the Nav commandline tool to integrate Graphviz
functionality for generating diagrams within the tool itself, eliminating
the need to pipe output to an external Graphviz tool like dot. While the
long-term goal is to create a web application for navigating code with
diagram display, the objective is to maintain Nav as a commandline tool.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>