Add: dot_parser to validate generated dot files
This patch introduces a small Flex - Bison dot parser.
I'm not claiming it is essential, but it came easy from a work I did
before, and I think it to be useful for testing the nav application.
It is written in C which make nav use CGO at some point. For now,
just the flex - bison - c code of the parser.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>