Partial refactor sql
The SQL component used to be this clunky, old-school set of functions
written in C style. But with this patch, I've gone ahead and given it
a total makeover by abstracting it with interfaces and receivers. The
result is a shiny new version that makes it way easier to make changes
to the persistence code down the line.
You could even add in new components, like non-SQL databases, just by
sticking to the interface.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>
Suggested-by: Maurizio Papini <mpapini@redhat.com>