feat: Add support for C-style switchable debug messages
This commit adds the ability to turn debug messages on and off through a
switch. When the switch is turned off, no code is inserted, but when it is
turned on, code is generated and messages are printed. The benefit of this
implementation is that messages can be left in place without impacting
performance or producing output when the switch is turned off.

Please see
https://carminatialessandro.blogspot.com/2023/04/c-styel-print-debug-in-golang.html
for additional references.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>